Converting to Node
Using Node.js as the main REST-based API component. This is being done to make it easier to script API functions. Changes also made to the README to reflect this. The writeToDb.py file remains as it is used to enter data into the database and has nothing to do with the API itself.